refactor(config): split helper tests for GH727
